add sensorType, some small refinement

+++ EmulatedSensor/sensor.sh
@@ -1,24 +1,56 @@
 export BASE=`dirname $0`
+cd $BASE
 
-if [ $# -lt 1 ]
-then
-        echo "Usage : $0 <command> start|stop|change-rate|connect-console|connect-mqtt"
-        echo "Command:"
+listCommand() {
+
         echo "  start"
         echo "  stop"
         echo "  change-rate <rate>"
         echo "  connect-console"
         echo "  connect-mqtt <broker> [topic]"
+}
+
+if [ $# -lt 1 ]
+then
+        echo "Usage : $0 <command> start|stop|change-rate|connect-console|connect-mqtt"
+        echo "Command:"
+        listCommand
         exit
 fi
 
 case "$1" in
+ list-commands)
+   listCommand
+ ;;
  start)
+   if [ -f $BASE/sensor.pid ]
+   then
+     PID=`cat sensor.pid`
+     if ps -p "$PID" > /dev/null
+     then
+       echo "Sensor started already. PID: $PID"
+       exit 1
+     fi
+   fi
+ 
    java -jar EmulatedSensor-1.0-SNAPSHOT.jar  > sensor.out 2>&1 &
-   echo $!>sensor.pid
+   pid=$!
+   if [ ! -z "$pid" ]
+   then
+     echo "Sensor started. PID: $pid"
+     echo $pid > sensor.pid
+   else
+     echo "Failed to start sensor. Please check sensor.log for more detail."
+   fi
+   
    ;;
    
  stop)
+   if [ ! -f $BASE/sensor.pid ]
+   then
+     echo "Sensor seems to be stopped, nothing to do."
+     exit 1
+   fi
    kill `cat $BASE/sensor.pid`
    rm $BASE/sensor.pid
    echo "Sensor stopped sucessfully!"
@@ -44,9 +76,9 @@ case "$1" in
    
  connect-mqtt)
    sed -i 's#platform=.*#platform=teit.sensor.MQTT.MQTTOutput#' $BASE/sensor.conf
-   sed -i 's#platform.mqtt.url=.*#platform.mqtt.url='$BROKER'#' $BASE/sensor.conf
+   sed -i 's#platform.mqtt.url=.*#platform.mqtt.url='$2'#' $BASE/sensor.conf
    if [ ! -z "$3" ]; then
-     sed -i 's#platform.mqtt.topic=.*#platform.mqtt.topic='$TOPIC'#' $BASE/sensor.conf
+     sed -i 's#platform.mqtt.topic=.*#platform.mqtt.topic='$3'#' $BASE/sensor.conf
    fi
    touch sensor.conf
    ;;
